Move applyChartTooltips fn into separate namespace
