+++ frontend/src/metabase/visualizations/lib/apply_axis.js
